Split utilities for @testing-library/react and react-test-renderer
