+++ src/components/ProjectScreen/tests/ChooseProject.test.tsx
@@ -3,10 +3,10 @@ import renderer from "react-test-renderer";
 
 import "tests/reactI18nextMock";
 
-import { Project } from "api/models";
+import { type Project } from "api/models";
 import ChooseProject from "components/ProjectScreen/ChooseProject";
 import { newProject } from "types/project";
-import { testInstanceHasText } from "utilities/testUtilities";
+import { testInstanceHasText } from "utilities/testRendererUtilities";
 import { randomIntString } from "utilities/utilities";
 
 jest.mock("backend", () => ({
