+++ src/goals/CharacterInventory/CharInv/CharacterDetail/tests/index.test.tsx
@@ -1,5 +1,5 @@
 import { Provider } from "react-redux";
-import { ReactTestRenderer, act, create } from "react-test-renderer";
+import { type ReactTestRenderer, act, create } from "react-test-renderer";
 import configureMockStore from "redux-mock-store";
 
 import "tests/reactI18nextMock";
@@ -12,8 +12,8 @@ import {
 } from "goals/CharacterInventory/CharInv/CharacterDetail/FindAndReplace";
 import CharacterReplaceDialog from "goals/CharacterInventory/CharInv/CharacterDetail/FindAndReplace/CharacterReplaceDialog";
 import { defaultState } from "goals/CharacterInventory/Redux/CharacterInventoryReduxTypes";
-import { StoreState } from "types";
-import { testInstanceHasText } from "utilities/testUtilities";
+import { type StoreState } from "types";
+import { testInstanceHasText } from "utilities/testRendererUtilities";
 
 // Dialog uses portals, which are not supported in react-test-renderer.
 jest.mock("@mui/material", () => {
